fixed wrong words
