+++ data/data_proceed.py
@@ -2,8 +2,8 @@ import mysql.connector
 import json
 
 mydb = mysql.connector.connect(
-  host = "localhost",
-  user = "root",
+  host = "0.0.0.0",
+  user = "lucas",
   password = "00000000",
 )
 mycursor1 = mydb.cursor()
